requirements.txt
polars==0.20.26
requests==2.31.0
pyodbc==4.0.39
python-dotenv==1.0.0
urllib3>=1.26.0